requirements.txt
fastapi
uvicorn
gradio
requests
langchain
langchain-community
langchain-google-genai
openai
anthropic
python-dotenv
